Evren v2 + soyut düşünce altyapısı: semboller, özyinelemeli düşünce, ayrık dil, ölüm farkındalığı, yazıt sistemi
Universe v2: ecology layer, hidden physics (radiation, magnetic, underground, soil pH),
expanded hypothesis system (40 features, 20 outcomes, 5 comparators), terrain/chemistry integration.

Abstract thinking infrastructure (4 phases):
- Phase 1: VQ symbol codebook - discretizes bottleneck concepts into composable symbols
- Phase 2: Recursive thought - GRU sandbox thinking with depth gate (genome-controlled)
- Phase 3: Discrete token language - grounded vocabulary with speaker/listener learning
- Phase 4: Mortality self-model - concept-space death awareness (no hardcoded semantics)
- Inscription system: agents write token sequences onto structures for persistent knowledge transfer

8 new genome genes (FIXED_GENE_COUNT 31→39): n_symbols, n_symbol_slots, symbol_lr,
depth_gate_threshold, vocab_active, utterance_length, listen_weight, mortality_sensitivity.

Design principle: "No semantic hardcoding above the physics layer" - all meaning is learned.

+++ world/grid.py
@@ -14,7 +14,7 @@ class Grid:
     def __init__(self, cfg: Config, rng: np.random.Generator):
         self.w = cfg.world_width
         self.h = cfg.world_height
-        self.cells = np.zeros((self.w, self.h, NUM_CHANNELS), dtype=np.float64)
+        self.cells = np.zeros((self.w, self.h, NUM_CHANNELS), dtype=np.float32)
         self._rng = rng
         self._cfg = cfg
